% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{catchdata_transformation}
\alias{catchdata_transformation}
\title{Transformation of basic catch data}
\usage{
catchdata_transformation(data)
}
\arguments{
\item{data}{The basic catch data frame.}
}
\description{
This function is used to transform stock-based fisheries catch data into a suitable format for the clustering procedure and cluster validation of the FleetSegmentation package.
It uses the share of the species on the total catch of each ship instead of the amount (or other seafood) of fish caught.
See example_catchdata for a example data frame format. Catch needs to be given in kilograms.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
}
\keyword{data}
\keyword{transformation}
