% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{cluster_assemblages_MDS}
\alias{cluster_assemblages_MDS}
\title{Multi-dimensional scaling (MDS) of the assemblage-based cluster catches}
\usage{
cluster_assemblages_MDS(
  data,
  catchdata,
  clustering,
  interactive = F,
  GoF = T,
  distance = "jaccard",
  dim = 2
)
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{catchdata}{The transformed catchdata created with catchdata_transformation()}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{GoF}{Display goodness of fit in the MDS plot. Defaults to TRUE}

\item{distance}{The distance measure used. Defaults to modified (metric conversion) Bray-Curtis distance distance. CAUTION! The clustering approach for the fleet segmentation is designed to work with modified (metric-converted) Bray-Curtis distance and the average linkage method! Changing either of them is not advised!}

\item{dim}{The dimensions of the MDS. Use `2` for a 2-dimensional, classic MDS and `3` for a 3-dimensional MDS.}
}
\description{
This is function creates an MDS of the assemblage-based catches of the clustering result. The MDS can be either 2-dimensional (which is the default setting), or 3-dimensional.
3-dimensional MDS are harder to interpret, but due to the nature of compositional catch data, 2-dimensional MDS often have a poor goodness of fit (GoF) and have to be treated with caution.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
cluster_assemblages_MDS(data =data, catchdata = catchdata,clustering = clustering, GoF=TRUE)
}
\keyword{MDS}
\keyword{clustering}
