% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{clustering_MDS}
\alias{clustering_MDS}
\title{Multi-dimensional scaling (MDS) of the clustering}
\usage{
clustering_MDS(catchdata, clustering, dim = 2, GoF = T)
}
\arguments{
\item{catchdata}{The transformed catchdata created with catchdata_transformation()}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{dim}{The dimensions of the MDS. Use `2` for a 2-dimensional, classic MDS and `3` for a 3-dimensional MDS.}

\item{GoF}{Display goodness of fit in the MDS plot. Defaults to TRUE}
}
\description{
This is function creates an MDS of the clustering result. The MDS can be either 2-dimensional (which is the default setting), or 3-dimensional.
3-dimensional MDS are harder to interpret, but due to the nature of compositional catch data, 2-dimensional MDS often have a poor goodness of fit (GoF) and have to be treated with caution.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
clustering_MDS(catchdata = catchdata,clustering = clustering, GoF=TRUE)
clustering_MDS(catchdata = catchdata,clustering = clustering,dim = 3)
}
\keyword{MDS}
\keyword{clustering}
