% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{clustering_plotgrid}
\alias{clustering_plotgrid}
\title{Plotgrid of number of ships, ship length, catch of single ships and total catch of ships in clusters}
\usage{
clustering_plotgrid(data, clustering, shiplength, subset = NULL)
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{shiplength}{A data frame containing the length of the ships clustered.}

\item{subset}{Display only a subset of clusters in plot. Can be a single number or a vector of numbers.}
}
\description{
This is function creates an overview plot grid of
1) A barplot of the number of ships in each cluster
2) A mixed dot- and boxplot of the length of ships in the clusters
3) A mixed dot- and boxplot of the catch of single ships in the clusters.
4) A barplot of the total catch of all ships in each cluster.
Boxplots will only be drawn for clusters containing more than 5 ships.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
clustering_plotgrid(data = stockdata,clustering = clustering,shiplength =example_lengthdata)
}
\keyword{clustering}
\keyword{grid}
\keyword{plot}
