% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{clustering_stockshares_table}
\alias{clustering_stockshares_table}
\title{Tabelize stock shares of clusters}
\usage{
clustering_stockshares_table(data, clustering, style = "basic")
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{style}{The style, in which the result is printed. Defaults to "basic". "html" produces an html-table.}
}
\description{
This function creates a table with the average shares of the stocks on the catch of each clusters vessel. The result is a data frame, which can be printed or stored as an object.
Alternatively, an html-table can be created by setting style = "html".
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
clustering_stockshares_table(data = stockdata,clustering = clustering)
clustering_stockshares_table(data = stockdata,clustering = clustering,style = "html")
segments_stockshares <- clustering_stockshares_table(data = stockdata,clustering = clustering)
}
\keyword{clustering}
\keyword{stockshares}
\keyword{table}
