% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_catchdata.R
\docType{data}
\name{example_catchdata}
\alias{example_catchdata}
\title{Catch data of the example ships from the FleetSegmentation package}
\format{
A data frame with 287 rows and 4 variables:
\describe{
  \item{ship_ID}{ship identification number}
  \item{species}{(Three-letter ICES species code}
  \item{fao_area}{FAO area code}
  \item{landings}{catch [kg]}
}
}
\usage{
example_catchdata
}
\description{
A dataset containing the total catch per species and FAO area of the 71 example ships from the FleetSegmentation package.
Created by Erik Sulanke, Thuenen-Institute for Sea Fisheries, Bremerhaven, Germany, in September 2020, modified in March 2021.
}
\keyword{datasets}
