% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_lengthdata.R
\docType{data}
\name{example_lengthdata}
\alias{example_lengthdata}
\title{Length data of the example ships from the FleetSegmentation package}
\format{
A data frame with 71 rows and 2 variables:
\describe{
  \item{ship_ID}{ship identification number}
  \item{shiplength}{length of the ship [m]}
}
}
\usage{
example_lengthdata
}
\description{
A dataset containing the ship length over all in meters of the 71 example ships from the FleetSegmentation package.
Created by Erik Sulanke, Thuenen-Institute for Sea Fisheries, Bremerhaven, Germany, in September 2020.
}
\keyword{datasets}
