% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{HHI_plot}
\alias{HHI_plot}
\title{Mixed box- and dotplot of Herfindahl-Hirschmann Index (HHI) of catch of single ships and overall catch of clusters.}
\usage{
HHI_plot(data, clustering)
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}
}
\description{
This function creates a mixed box- and dotplot of the HHI of the catch of 1) single ships and 2) clusters. The HHI is equivalent to the Simpson Index,
it is a measure of degree of concentration/consolidation. In case of fisheries catch data, it indicates, whether a fishery is of rather targeted or mixed nature.
The HHI can take values between 0 and 1. An index value below 0.25 indicates a highly mixed fishery, between 0.25 and 0.5 a rather mixed fishery, between 0.5 and 0.75 a rather targeted fishery and above 0.75 a highly targeted fishery.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
HHI_plot(data = stockdata,clustering = clustering)
}
\keyword{HHI}
\keyword{clustering}
\keyword{plot}
