% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{HHI_table}
\alias{HHI_table}
\title{Table of HHI of  overall catch of clusters.}
\usage{
HHI_table(data, clustering, style = "basic")
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{style}{The style, in which the result is printed. Defaults to "basic". "html" produces an html-table.}
}
\description{
This function creates an overview table of the Herfindahl-Hirschmann Index (HHI) of the catch clusters. The HHI is equivalent to the Simpson Index,
it is a measure of degree of concentration/consolidation. In case of fisheries catch data, it indicates, whether a fishery is of rather targeted or mixed nature.
The HHI can take values between 0 and 1. An index value below 0.25 indicates a highly mixed fishery, between 0.25 and 0.5 a rather mixed fishery, between 0.5 and 0.75 a rather targeted fishery and above 0.75 a highly targeted fishery.
The result of the function is a data frame, which can be printed or stored as an object.
Alternatively, an html-table can be created by setting style = "html".
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
HHI_table(data = stockdata,clustering = clustering)
}
\keyword{HHI}
\keyword{clustering}
\keyword{plot}
