% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{numberclust_dendrogram}
\alias{numberclust_dendrogram}
\title{Plot dendrogram to identify number of clusters}
\usage{
numberclust_dendrogram(
  catchdata,
  distance = "jaccard",
  method = "average",
  dend_method = "basic",
  dend_cut = 0.75,
  range_min = 0.5,
  range_max = 1
)
}
\arguments{
\item{catchdata}{The transformed catchdata created with catchdata_transformation()}

\item{distance}{The distance measure used. Defaults to modified (metric-converted) Bray-Curtis distance. CAUTION! The clustering approach for the fleet segmentation is designed to work with modified (metric-converted) Bray-Curtis distance and the average linkage method! Changing either of them is not advised!}

\item{method}{The link function used. Defaults to average linkage. CAUTION! The clustering approach for the fleet segmentation is designed to work with modified (metric-converted) Bray-Curtis distance and the average linkage method! Changing either of them is not advised!}

\item{dend_method}{The style of the plotted dendrogram. "basic" returns a blank dendrogram, "range" a dendrogram with the recommended cutting heights depicted, "cut" enables the user to cut the dendrogram at a height of his choice. The resulting number of clusters will be shown in the plot.}

\item{dend_cut}{The height used to cut the dendrogram. Defaults to 0.75, which showed to be the appropriate cutting height for various fleet data sets.}

\item{range_min}{The lower border of the cutting range set for a dendrogram with the "range"- method}

\item{range_max}{The upper border of the cutting range set for a dendrogram with the "range"- method}
}
\description{
This function plots a dendrogram of the clustering. The dendrogram can either be displayed without any further features
or with the recommended cutting heights (setting dend_method to "range") or a cut at a defined linkage distance (setting dend_method to "cut").
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
numberclust_dendrogram(catchdata=catchdata) ### basic dendrogram
numberclust_dendrogram(catchdata=catchdata, dend_method= "range")
### dendrogram with range of good cutting heights
numberclust_dendrogram(catchdata=catchdata, dend_method= "cut", dend_cut = 0.75)
### dendrogram with cut at a linkage distance of 0.75 and the number of resulting clusters
}
\keyword{clusters}
\keyword{dendrogram}
\keyword{number}
\keyword{of}
\keyword{plot}
