% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{numberclust_table}
\alias{numberclust_table}
\title{Compute most important indices for best number of clusters in table format}
\usage{
numberclust_table(
  catchdata,
  max_clusternumber = ifelse(nrow(catchdata) <= 15, (nrow(catchdata) - 1), 15),
  style = "basic",
  distance = "jaccard",
  method = "average"
)
}
\arguments{
\item{catchdata}{The transformed catchdata created with catchdata_transformation()}

\item{max_clusternumber}{The maximum number of clusters to be expected. Defaults to 1 less than the number of ships in the catchdata-frame, up to a maximum of 15.}

\item{style}{The output style, defaults to `basic`, which prints a data frame in the console and can be stored. For a html-version, use `html`}

\item{distance}{The distance measure used. Defaults to modified (metric conversion) Bray-Curtis distance distance. CAUTION! The clustering approach for the fleet segmentation is designed to work with modified (metric-converted) Bray-Curtis distance and the average linkage method! Changing either of them is not advised!}

\item{method}{The link function used. Defaults to average linkage. CAUTION! The clustering approach for the fleet segmentation is designed to work with modified (metric-converted) Bray-Curtis distance and the average linkage method! Changing either of them is not advised!}
}
\description{
The fleet segmentation package uses the the average silhouettes, the Mantel test, the Davis-Bouldin index,
the SD-index and the Calinski-Harabasz index. This function gives the values of those indices for the given maximal number of clusters in a table, which can be printed in a basic or html format or stored as a data frame.
A modified (metric converted) Bray-Curtis distance matrix is computed from the input data, the clustering is performed as a hierarchical agglomerative clustering (HAC) using the average linkage link function.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
numberclust_table(catchdata = catchdata,max_clusternumber = 15)
numberclust_table(catchdata = catchdata,max_clusternumber = 15,style = "html")
optclust_table <- numberclust_table(catchdata = catchdata,max_clusternumber = 15)
}
\keyword{clusters}
\keyword{number}
\keyword{of}
\keyword{table}
