% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{segmentation_datapreparation}
\alias{segmentation_datapreparation}
\title{Function to automatically separate a fleet data frame by gear class and assign ICES and ICCAT stocks.}
\usage{
segmentation_datapreparation(
  fleetdata,
  vessel_ID,
  shiplength,
  gear,
  species,
  area,
  catch,
  reduce = T,
  auto.generate = T,
  threshold.auto.generate = 100,
  min.share = 0
)
}
\arguments{
\item{fleetdata}{The underlying fleet data frame. With the following arguments, the columns are specified}

\item{vessel_ID}{The unquoted name of the column containing the vessel identifier}

\item{shiplength}{The unquoted name of the column containing the length of the vessel in m}

\item{gear}{The unquoted name of the column containing the code for the gear class}

\item{species}{The unquoted name of the column containing the three-digit code for the species caught}

\item{area}{The unquoted name of the column containing the code for the area in ICES/FAO coding (e.g., 27.4.a or GSA 16)}

\item{catch}{The unquoted name of the column containing the catch in kg}

\item{reduce}{Indicates, whether or not the resulting data frame is reduced to Ship ID, ICES stock and catch weight. Defaults to TRUE.
Turned to FALSE, the resulting data frame will resemble the original data with the ICES-Stock column added. This format is used for control and correction purposes,
but not for further calculations.}

\item{auto.generate}{Indicates whether or stocks should be automatically generated for
a) species, which are not assessed in ICES-Stocks or
b) assessed by ICES, but caught out of stock-managed areas.
The automatically generated stocks comprise the species name and the FAO area.
The relevant quantity is defined by the argument threshold.auto.generate}

\item{threshold.auto.generate}{Threshold of automatic generation of ICES stocks. Only relevant if auto.generate = T. Defaults to 100.}

\item{min.share}{The minimal share a stock has to have on at least one vessels catch to be included in the stock dataframe. Defaults to 0, i.e. every stock is retained by default.}
}
\description{
This function separates given fleet data by gear class and creates separate, conveniently named and ready-to-use data frames.
It also assigns ICES stocks to catch data based on the caught species and the FAO fishing area where it was caught.
Due to limits in available data, especially on a fine spatial resolution, not all ICES-stocks can be taken into account. Therefore, stocks of Nephrops norvegicus
and Ammodytes spp. are based on EU definitions, not on ICES definitions. Additionally, the function automatically creates stocks based on species name and FAO area for species
which are caught in larger quantities (> 100 kg) but are not part of defined ICES-stock. This feature can be shut off if the user decides not to apply the procedure and
use only defined ICES-stocks.
}
\examples{
segmentation_datapreparation(data = fleet_exmpl,
                              vessel_ID = ship_ID,
                              shiplength = loa,
                              gear = main_gear,
                              species = spec_code,
                              area = fao_area,
                              catch = landkg)
}
\keyword{data}
\keyword{preparation}
