% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{shiplength_plot}
\alias{shiplength_plot}
\title{Plot length of ships in clusters}
\usage{
shiplength_plot(clustering, shiplength)
}
\arguments{
\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{shiplength}{A data frame containing the length of the ships clustered.}
}
\description{
This is function creates an overview mixed dot- and boxplot of the length of the ships in each cluster. The length can be given in cm or m, the function will auto-transform from cm to m.
A boxplot will only be drawn for clusters containing more than 5 ships.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
shiplength_plot(clustering = clustering,shiplength = example_lengthdata)
}
\keyword{clustering}
\keyword{plot}
\keyword{shiplength}
