% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{single_cluster_stockshares}
\alias{single_cluster_stockshares}
\title{Plot stock shares of clusters}
\usage{
single_cluster_stockshares(
  data,
  clustering,
  min_share = 5,
  cluster.number,
  label_wrap = 15
)
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{min_share}{The minimum average percentage share a stock has to have to be labelled in the plot. Defaults to 5\%.}

\item{cluster.number}{The number of the cluster of which vessel stock shares should be displayed.}

\item{label_wrap}{Indicates the number of characters per line before a line break in the stock labels. Defaults to 15.}
}
\description{
This is function creates an overview barplot of the average shares of stocks on the catch of all the vessels in one cluster.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
single_cluster_stockshares(data = stockdata,clustering = clustering, cluster.number=1)
}
\keyword{clustering}
\keyword{plot}
\keyword{stockshares}
