% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FleetSegmentation.R
\name{singleship_catch_plot}
\alias{singleship_catch_plot}
\title{Plot of catch of single ships in clusters}
\usage{
singleship_catch_plot(data, clustering, subset = NULL)
}
\arguments{
\item{data}{The original, untransformed data that was used for the clustering.}

\item{clustering}{The result of the clustering procedure, stored as a data frame.}

\item{subset}{Display only a subset of clusters in plot. Can be a single number or a vector of numbers.}
}
\description{
This is function creates an overview mixed box- and dotplot of the catch of single ships in each cluster. A boxplot will only be drawn for clusters containing more than 5 ships.
}
\examples{
data <- example_catchdata
stockdata <- assign_stocks(data=data)
catchdata <- catchdata_transformation(data = stockdata)
clustering <- segmentation_clustering(catchdata = catchdata,n_cluster = 6)
singleship_catch_plot(data = stockdata,clustering = clustering)
}
\keyword{catch}
\keyword{clustering}
\keyword{plot}
